requirements.txt
django_bootstrap5
django
mysqlclient
django-crispy-forms
pillow
crispy-boostrap5